requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
pydantic==1.10.11
pytest==7.4.0
python-dotenv==1.0.0
websockets==11.0.3
aiortc==1.6.0
PyYAML==6.0
